# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DZ Bus Tracker API",
    "description": "Real-time bus tracking for Algeria.\n\n* JSON uses snake_case; timestamps are RFC 3339 UTC.\n* Errors are RFC 9457 problem documents (`application/problem+json`) with a stable `code`, localized in French (default), Arabic or English via `Accept-Language`.\n* Lists use cursor pagination: `?limit=` (1–100) and `?cursor=` from `next_cursor`.\n* Cacheable GETs return a weak `ETag`; send `If-None-Match` to get `304`.\n* Rate limits are reported in `RateLimit-*` headers; `429` carries `Retry-After`.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stops": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "List stops (newest first). Active stops only, unless the caller administers stops.",
        "operationId": "list_stops",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Part of the name (2–100 characters, case-insensitive).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "wilaya",
            "in": "query",
            "description": "Wilaya (exact, case-insensitive).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commune",
            "in": "query",
            "description": "Commune (exact, case-insensitive).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "line_id",
            "in": "query",
            "description": "Only stops served by this line.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "Activity filter, for stop administrators only (`403` otherwise).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of stops",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_StopDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "401": {
            "description": "`is_active` filter without being signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "`is_active` filter without `stop:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filters (e.g. `q` shorter than 2 characters)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "stops"
        ],
        "summary": "Create a stop (audited).",
        "operationId": "create_stop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Stop created (`Location` points to it)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `stop:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/stops/nearby": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "Active stops around a point, nearest first (not paginated).",
        "operationId": "nearby_stops",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "Latitude of the point (degrees).",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "description": "Longitude of the point (degrees).",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radius_m",
            "in": "query",
            "description": "Search radius in metres (10–5000, default 500).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of stops (1–50, default 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stops within the radius, with their distance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemList_NearbyStopDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "422": {
            "description": "Missing or out-of-range coordinates, radius or limit",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/stops/{id}": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "One stop. Inactive stops are only visible to stop administrators.",
        "operationId": "get_stop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stop id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stop",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "404": {
            "description": "No such (visible) stop",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "stops"
        ],
        "summary": "Delete a stop that no line serves (audited). Its photo is deleted in the background.",
        "operationId": "delete_stop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stop id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `stop:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such stop",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "A line still serves the stop (`stop_in_use`), or its photo changed concurrently (`stale_state`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "stops"
        ],
        "summary": "Update a stop partially (audited). Moving a stop recomputes the distances of its lines.",
        "operationId": "update_stop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stop id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated stop",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `stop:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such stop",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Lines kept gaining the moved stop concurrently (`stale_state`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/stops/{id}/lines": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "Lines serving a stop, ordered by code (inactive lines for line administrators only).",
        "operationId": "stop_lines",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stop id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The lines serving the stop",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemList_LineDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "404": {
            "description": "No such (visible) stop",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStopRequest": {
        "type": "object",
        "description": "A new stop.",
        "required": [
          "name",
          "location"
        ],
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ],
            "description": "At most 255 characters."
          },
          "commune": {
            "type": [
              "string",
              "null"
            ],
            "description": "At most 100 characters."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "At most 2000 characters."
          },
          "features": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "At most 20 tags of 1–40 characters `[a-z0-9_]` (lower-cased)."
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Default `true`."
          },
          "location": {
            "$ref": "#/components/schemas/LatLng"
          },
          "name": {
            "type": "string",
            "description": "1–100 characters.",
            "example": "Place des Martyrs"
          },
          "wilaya": {
            "type": [
              "string",
              "null"
            ],
            "description": "At most 100 characters."
          }
        },
        "additionalProperties": false
      },
...
      "Paginated_StopDto": {
        "type": "object",
        "description": "A page of results. Pass `next_cursor` as `cursor` to fetch the next page.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A bus stop.",
              "required": [
                "id",
                "name",
                "location",
                "address",
                "wilaya",
                "commune",
                "description",
                "features",
                "is_active",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "address": {
                  "type": "string"
                },
                "commune": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "description": {
                  "type": "string"
                },
                "features": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Tags such as `shelter`, `bench`, `wheelchair_access`."
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "is_active": {
                  "type": "boolean",
                  "description": "Inactive stops are only visible to stop administrators."
                },
                "location": {
                  "$ref": "#/components/schemas/LatLng"
                },
                "name": {
                  "type": "string"
                },
                "photo_url": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Presigned download URL of the photo (stable for an hour); `null` without photo or when\nfile storage is not available."
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "wilaya": {
                  "type": "string"
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ProblemDocument": {
        "type": "object",
        "description": "RFC 9457 problem document returned by every error response.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            }
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "URI identifying the problem type (`urn:dzbus:problem:<code>`)."
          }
        }
      },
...
      "StopDto": {
        "type": "object",
        "description": "A bus stop.",
        "required": [
          "id",
          "name",
          "location",
          "address",
          "wilaya",
          "commune",
          "description",
          "features",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "commune": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags such as `shelter`, `bench`, `wheelchair_access`."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean",
            "description": "Inactive stops are only visible to stop administrators."
          },
          "location": {
            "$ref": "#/components/schemas/LatLng"
          },
          "name": {
            "type": "string"
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Presigned download URL of the photo (stable for an hour); `null` without photo or when\nfile storage is not available."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "wilaya": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Machine-to-machine key (`dzk_…`) with explicit scopes."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from /auth/login or /auth/refresh (EdDSA)."
      }
    }
  },
...
}